Move miri failure to a more appropriate module
